Assignment 3, Task 9
Last week, there was an Extra Credit task to use a UISlider to show the
history of the currently-being-played game. This week we’re going to
make showing the history a Required Task, but instead of using a
slider, you must invent yet another new MVC which displays the history
in a UITextView and which is segued to inside a UINavigationController.
It should show the cards involved in every match or mismatch as well as
how many points were earned or lost as a result (you are already
showing this information in a UILabel for each card choosing, so this
should be rather straightforward to implement). Add a bar button item
“History” on the right side of the navigation bar which performs the
push segue. This feature must work for both the Playing Card game and
the Set game.